Position instances can be packed/unpacked to reduce network traffic. Made positions in code shorts rather than ints. REfresh on CLI client someytimes doesnt show spawn, theading/network issue?

+++ src/gaia/server/Constants.java
@@ -6,16 +6,16 @@ package gaia.server;
 public class Constants 
 {
 	/** The size of the world (tiles). */
-	public static int WORLD_SIZE = 1024;
+	public static short WORLD_SIZE = 1024;
 	
 	/** The size of a chunk (tiles). */
-	public static int WORLD_CHUNK_SIZE = 32;
+	public static short WORLD_CHUNK_SIZE = 32;
 	
 	/** The number of chunks per axis. */
-	public static int WORLD_CHUNKS_PER_AXIS = WORLD_SIZE / WORLD_CHUNK_SIZE;
+	public static short WORLD_CHUNKS_PER_AXIS = (short) (WORLD_SIZE / WORLD_CHUNK_SIZE);
 	
 	/** The distance (in chunks) from a chunk that is regarded as being in the chunks vicinity. */
-	public static int WORLD_CHUNK_VICINITY_RANGE = 1;
+	public static short WORLD_CHUNK_VICINITY_RANGE = 1;
 	
 	/** The server clock rate (ms). */
 	public static long SERVER_CLOCK_RATE = 125l;
